requirements.txt
opencv-python
numpy
cupy-cuda11x
timm
Pillow